Move fetchmode.h to utils
