+++ include/ace/utils/fetchmode.h
@@ -2,8 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#ifndef _ACE_MANAGERS_VIEWPORT_FETCHMODE_H_
-#define _ACE_MANAGERS_VIEWPORT_FETCHMODE_H_
+#ifndef _ACE_UTILS_FETCHMODE_H_
+#define _ACE_UTILS_FETCHMODE_H_
 
 #include <ace/types.h>
 #include <ace/utils/extview.h>
@@ -121,4 +121,4 @@ static inline LONG fetchModeCalcBplOffsetX(const tVPort *pVPort, UWORD uwScrollX
 	return lBplAddX;
 }
 
-#endif // _ACE_MANAGERS_VIEWPORT_FETCHMODE_H_
+#endif // _ACE_UTILS_FETCHMODE_H_
