+++ src/ace/managers/viewport/scrollbuffer.c
@@ -4,8 +4,8 @@
 
 #include <ace/managers/viewport/scrollbuffer.h>
 #include <ace/utils/tag.h>
+#include <ace/utils/fetchmode.h>
 #include <ace/generic/screen.h> // Has the look up table for the COPPER_X_WAIT values.
-#include <ace/managers/viewport/fetchmode.h>
 #include <limits.h>
 
 static UWORD nearestPowerOf2(UWORD uwVal) {
